backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
PyMySQL>=1.1,<2.0
bcrypt>=4.2,<5.0
PyJWT>=2.9,<3.0
